Fill inner and outer track geometries with a helix approximation.
Inner geometry will be filled at the first silicon layer, regardless
of whether it was used in the fit or not.  This allows us to express
the helix with a sufficiently constant magnetic field.

The outer geometry is a gross approximation at the last sTGC layer.

... bit premature to check in.  code compiles, runs, calculates the
helix parameters.  But does not yet fill the geometry.